+++ backend/app/models/__init__.py
@@ -1,2 +1,3 @@
 from .core import *
+from .releases import *
 from . import projects
